Edit: Content classes

+++ src/main/java/com/example/Netflix/Content/GenreBridge/GenreAssignmentRequest.java
@@ -3,14 +3,14 @@ package com.example.Netflix.Content.GenreBridge;
 import java.util.List;
 
 public class GenreAssignmentRequest {
-    private Long contentId;
+    private Integer contentId;
     private List<Long> genreIds;
 
     public Integer getContentId() {
         return contentId;
     }
 
-    public void setContentId(Long contentId) {
+    public void setContentId(Integer contentId) {
         this.contentId = contentId;
     }
